% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_from_data_list.R
\name{df_from_data_list}
\alias{df_from_data_list}
\title{Get and Validate a Data Subset for cross-cohort Analysis}
\usage{
df_from_data_list(model, target_items, target_groups)
}
\arguments{
\item{model}{A `harm_model` object representing model structure}

\item{target_items}{A character vector of the item column names to select.}

\item{target_groups}{A character vector of the cohort names to select.
that identifies the groups. Defaults to "cohort".}
}
\value{
A dataframe containing only the rows for 'target_groups' and
  the columns for 'group_col' and 'items'.
}
\description{
This function subsets a dataframe.
It validates that the specified groups and item columns exist before returning
a filtered dataframe.
}
