% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_analysis.R
\name{dif_analysis}
\alias{dif_analysis}
\title{Run Exploratory DIF Analysis}
\usage{
dif_analysis(model, target_items, target_cohorts)
}
\arguments{
\item{model}{A `harm_model` object representing model structure}

\item{target_items}{A vector of names of targeted items}

\item{target_cohorts}{A vector of names of targeted cohorts}
}
\description{
This function run LASSO regularization test on every item of a linear model,
predict uniform DIF based on log-scale greatest gap method.
In each domain, the item with least DIF statistic will be selected as anchor,
the function will run MIMIC test on remaining items.
}
