% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_list_validate.R
\name{dif_list_validate}
\alias{dif_list_validate}
\title{Validate All Harmonization Inputs}
\usage{
dif_list_validate(model, unidif = NULL, nonunidif = NULL)
}
\arguments{
\item{model}{An object of class 'harm_model'.}

\item{unidif}{A list describing uniform DIF.}

\item{nonunidif}{A list describing non-uniform DIF.}
}
\description{
Runs a series of validation checks on the model and DIF lists
before starting harmonization.
}
