% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2nd.R
\name{fit2nd}
\alias{fit2nd}
\title{Fit a `harm_model` into Second Order Linear Models}
\usage{
fit2nd(model, selected_items, selected_cohorts, std = TRUE)
}
\arguments{
\item{model}{A `harm_model` object representing model structure.}

\item{selected_items}{A name or character vector of names of selected item(s).}

\item{selected_cohorts}{A name or A character vector of names of selected cohort(s).}

\item{std}{Logical. If true standardize general factor. Default to `TRUE`.}
}
\value{
a fit result, or a named list of fit results with `lavaan::sem`
}
\description{
Given full model information and selected items and groups, for each group,
fit the data with a individual second order general factor linear model.
Loadings are set free. Standardizing means and variances by default.
}
