% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_est_param.R
\name{get_est_param}
\alias{get_est_param}
\title{Extracts loadings, intercepts means, and residual variances for a
given list of items from a lavaan parameterEstimates data frame.}
\usage{
get_est_param(estimates_df, item_names)
}
\arguments{
\item{estimates_df}{A data.frame returned from `lavaan::parameterEstimates()`.}

\item{item_names}{A character vector of the observed variable names (items)
you want to extract parameters for.}
}
\value{
A data.frame with n rows (one for each item) and 3 columns:
        "loading", "mean", and "variance".
}
\description{
Extracts loadings, intercepts means, and residual variances for a
given list of items from a lavaan parameterEstimates data frame.
}
