% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmonization}
\alias{harmonization}
\title{Harmonize the Data Given Full Model Strucuture}
\usage{
harmonization(model, unidif = NULL, nonunidif = NULL)
}
\arguments{
\item{model}{A `harm_model` object representing model structure.}

\item{unidif}{A named list specifying the cohorts that each items is believed
to have uniform DIF. Default to be `NULL`.}

\item{nonunidif}{A named list specifying the cohorts that each items is believed
to have nonuniform DIF. Default to be `NULL`.}
}
\value{
A named of of fit results and the vector of all parameter values.
}
\description{
This function harmonize the data across cohorts if given `harm_model`, in order
of `harm_model$cohorts`. DIF items can be specified in advance. A named list of
fit results and a list of fixed parameters will be return
}
