% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_restriction.R
\name{lasso_restriction}
\alias{lasso_restriction}
\title{Run LASSO Regularization Test}
\usage{
lasso_restriction(model, target_items, target_cohorts)
}
\arguments{
\item{model}{A `harm_model` object representing model structure}

\item{target_items}{A vector of names of targeted items}

\item{target_cohorts}{A vector of names of targeted cohorts}
}
\value{
A table summarize prediction of DIF on targeted items
}
\description{
This function run LASSO regularization test on every item of a linear model,
predict uniform DIF based on log-scale greatest gap method, and print a plot
of regression coefficients and prediction of DIF.
}
\details{
Notice: non-uniform DIF causing poor fit
}
