% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvn_syntax.R
\name{lvn_syntax}
\alias{lvn_syntax}
\title{Generate `lavaan` Syntax of First Order Linear Model}
\usage{
lvn_syntax(
  model,
  selected_items,
  loadings = NULL,
  fix_lv_means = FALSE,
  fix_lv_variances = FALSE,
  fixed_intercept = 0,
  fixed_variance = 1,
  group_reg = NULL
)
}
\arguments{
\item{model}{A `harm_model` object representing model structure.}

\item{selected_items}{A character vector of names of selected items.}

\item{loadings}{A named numeric vector. Names are selected item names, values
are the loadings to fix. Items not in this vector will be freely estimated.
If `NULL` (default), all loadings are freely estimated.}

\item{fix_lv_means}{Logical.
If `TRUE`, latent variable intercepts are fixed (see `fixed_intercept`).
If `FALSE` (default), intercepts are freely estimated.}

\item{fix_lv_variances}{Logical.
If `TRUE`, latent variable variances are fixed (see `fixed_residual_var`).
If `FALSE` (default), variances are freely estimated.}

\item{fixed_intercept}{The numeric value to fix item intercepts to when
`fix_means = TRUE`. Default is `0.0`.}

\item{fixed_variance}{The numeric value to fix item residual variances
to when `fix_lv_variances = TRUE`. Default is `1.0`.}

\item{group_reg}{A character vector of item names.
If non-empty, items in it will add a regression to the group variable.
If `NULL` (default), skip this part.}
}
\value{
A character string of the formatted lavaan model syntax.
}
\description{
Given full model information, generate model syntax of a 1st order linear model
include only selected items.
}
