% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvn_syntax_g.R
\name{lvn_syntax_g}
\alias{lvn_syntax_g}
\title{Generate `lavaan` Syntax of Second Order General Factor Linear Model}
\usage{
lvn_syntax_g(
  model,
  selected_items,
  lv_loadings = NULL,
  ob_loadings = NULL,
  ob_mean = NULL,
  ob_var = NULL,
  std_g = FALSE,
  gf = "G"
)
}
\arguments{
\item{model}{A `harm_model` object representing model structure.}

\item{selected_items}{A character vector of names of selected items.}

\item{lv_loadings}{A named numeric vector. Names are selected domain factor names,
values are the loadings to fix. Domain factors not in this vector will be
freely estimated.
 If `NULL` (default), all loadings are freely estimated.}

\item{ob_loadings}{A named numeric vector. Names are selected item names,
values are the loadings to fix. Items not in this vector will be freely estimated.
 If `NULL` (default), all loadings are freely estimated.}

\item{ob_mean}{A named numeric vector. Names are selected item names,
values are the intercepts to fix. Items not in this vector will be freely estimated.
 If `NULL` (default), all intercepts are freely estimated.}

\item{ob_var}{A named numeric vector. Names are selected item names,
values are the variances to fix. Items not in this vector will be freely estimated.
 If `NULL` (default), all variances are freely estimated.}

\item{std_g}{Logical.
If `TRUE`, general factor intercept and variance, and esidual variance of latent
variables will be standardized.
If `FALSE` (default), these parameters are freely estimated.}

\item{gf}{A string representing the general factor's name. Default to `G`.}
}
\value{
A character string of the formatted lavaan model syntax.
}
\description{
Given full model information, generate model syntax of a 2st order linear model
include only selected items.
}
