% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimic.R
\name{mimic}
\alias{mimic}
\title{Run MIMIC Test}
\usage{
mimic(model, target_items, target_cohorts, anchors)
}
\arguments{
\item{model}{A `harm_model` object representing model structure}

\item{target_items}{A vector of names of targeted items}

\item{target_cohorts}{A vector of names of targeted cohorts}

\item{anchors}{a vector of names of anchor items}
}
\value{
A table summarize prediction of DIF on targeted items
}
\description{
This function run MIMIC(Multiple-Indicator, Multiple-Cause) test on selected
items of a linear model, excluding anchor items. This function reports z-scores
and p-scores of group regression given null hypothesis 0.
}
\details{
Notice: Can not detect DIF in factor loadings.
}
