% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_model.R
\name{specify_model}
\alias{specify_model}
\title{Define a Harmonization Model Specification}
\usage{
specify_model(
  cohorts,
  cohort_domain_test_map,
  covariate_lst,
  preferred_domain_order = c(),
  data,
  group_var = "cohort"
)
}
\arguments{
\item{cohorts}{A vector of names of all cohorts.}

\item{cohort_domain_test_map}{A map showing which tests belongs to each cohort's
each domain.}

\item{covariate_lst}{A list showing which ethnics info is presented in each cohort}

\item{preferred_domain_order}{A vector show preferred order of domains}

\item{data}{A named list, store the raw data by cohort.}

\item{group_var}{A string, the name of the grouping variable. Default to 'cohort'.}
}
\value{
An object of class 'harm_model'.
}
\description{
This function creates a 'harm_model' object that stores the model structure,
data, and settings for use in later functions.
}
