% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dif_list.R
\name{transform_dif_list}
\alias{transform_dif_list}
\title{Transforms an item-centric DIF list to a cohort-centric DIF list}
\usage{
transform_dif_list(dif_list)
}
\arguments{
\item{dif_list}{The input list}
}
\value{
A new list, named by cohort, where each element is a
        named vector of `item = group_id`.
}
\description{
Transforms an item-centric DIF list to a cohort-centric DIF list
}
